Create intial index.html, styles.css, basic layout
This is the first commit for the Landing Page Project. First build will
be a recreation of the given example without much creative input. This
first commit establishes the basic structure of the page, creating sections
for the header, hero section, main block, quote block, and footer. A few stock
images from the repository found at https://www.pexels.com/@padrinan/
were used as part of the layout. I got a little excited and began creating
the CSS for the page before creating this first commit. From here, the goal
will be to use flexbox to position the elements properly on the page.